+++ extensions/matter/src/components/Actions.tsx
@@ -1,15 +1,15 @@
 import { ActionPanel, Action, Icon, Toast, showToast } from "@raycast/api";
 import { useState } from "react";
-import { setFavorite } from "./matterApi";
+import { setFavorite } from "../matterApi";
 
 export function Actions(props: any) {
   const [isFavorited, setIsFavorited] = useState<boolean>(props.item.content.library.is_favorited);
+  const entryURL = "https://web.getmatter.com/entry/";
 
   async function favorite(isFavorited: boolean) {
     try {
       const res: any = await setFavorite(props.item.content.id, true);
       setIsFavorited(isFavorited);
-
       if (res.id && isFavorited) {
         showToast(Toast.Style.Success, "Success", "Article favorited");
       } else {
@@ -23,8 +23,12 @@ export function Actions(props: any) {
   return (
     <ActionPanel title={props.item.content.title}>
       <ActionPanel.Section>
-        {/* OPEN IN BROWSER */}
-        {props.item.content.url && <Action.OpenInBrowser url={props.item.content.url} />}
+        {/* OPEN IN MATTER */}
+        {props.item.content.article && props.item.content.id && (
+          <Action.OpenInBrowser url={entryURL + props.item.content.id} title="Open in Matter" />
+        )}
+        {/* View Original */}
+        {props.item.content.url && <Action.OpenInBrowser url={props.item.content.url} title="View Original" />}
         {/* COPY LINK */}
         {props.item.content.url && (
           <Action.CopyToClipboard
@@ -36,7 +40,7 @@ export function Actions(props: any) {
         {/* FAVORITE ARTICLE */}
         {props.item.content.id && !isFavorited && (
           <Action
-            title="Add to Favorites"
+            title="Add to favorites"
             icon={Icon.Star}
             onAction={async () => {
               await showToast(Toast.Style.Animated, "Loading");
@@ -48,7 +52,7 @@ export function Actions(props: any) {
         {/* UNFAVORITE ARTICLE */}
         {props.item.content.id && isFavorited && (
           <Action
-            title="Remove from Favorites"
+            title="Remove from favorites"
             icon={Icon.StarDisabled}
             onAction={async () => {
               await showToast(Toast.Style.Animated, "Loading");