+++ extensions/matter/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Matter Changelog
 
+## [🕹️ 1.1] - 2023-02-23
+
+- now you can open articles within Matter Web for a better reading experience - thanks to [Josh Knowles](https://github.com/joshknowles)
+- you can now filter your favorite articles
+
 ## [⭐️ Favorites] - 2023-01-08
 
 - now you can favorite articles (and remove them from favorites)
@@ -7,6 +12,7 @@
 
 ## [🎉 Initial release] - 2022-12-25
 
-Features: 
+Features:
+
 - showing saved for later articles
 - word count displayed